backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
python-dotenv
pydantic
requests
python-multipart
